+++ dogm/demo/CMakeLists.txt
@@ -1,0 +1,16 @@
+find_package(OpenCV)
+if (NOT OpenCV_FOUND)
+	set(OpenCV_INCLUDE_DIRS "" CACHE PATH "" )
+	set(OpenCV_LIBS "" CACHE FILEPATH "" )
+endif()
+
+add_executable(demo
+	main.cpp
+)
+
+target_include_directories(demo PUBLIC ${OpenCV_INCLUDE_DIRS} ${GLM_INCLUDE_DIR})
+
+target_link_libraries(demo
+	dogm
+	${OpenCV_LIBS}
+)
