+++ dogm/demo/main.cpp
@@ -23,6 +23,7 @@ SOFTWARE.
 */
 #include "dogm/dogm.h"
 #include "dogm/dogm_types.h"
+#include "simulator.h"
 
 #include <glm/glm.hpp>
 #include <opencv2/opencv.hpp>
@@ -43,52 +44,6 @@ SOFTWARE.
 using namespace std;
 
 #define PI 3.14159265358979323846f
-
-struct Vehicle
-{
-    Vehicle(const int width, const glm::vec2& pos, const glm::vec2& vel) : width(width), pos(pos), vel(vel) {}
-
-    void move(float dt) { pos += vel * dt; }
-
-    int width;
-    glm::vec2 pos;
-    glm::vec2 vel;
-};
-
-struct Simulator
-{
-    Simulator(int num_measurements) : num_measurements(num_measurements) {}
-
-    void addVehicle(const Vehicle& vehicle) { vehicles.push_back(vehicle); }
-
-    std::vector<std::vector<float>> update(int steps, float dt)
-    {
-        std::vector<std::vector<float>> measurements;
-
-        for (int i = 0; i < steps; i++)
-        {
-            std::vector<float> measurement(num_measurements, INFINITY);
-
-            for (auto& vehicle : vehicles)
-            {
-                vehicle.move(dt);
-
-                for (int i = 0; i < vehicle.width; i++)
-                {
-                    int index = static_cast<int>(vehicle.pos.x) + i;
-                    measurement[index] = vehicle.pos.y;
-                }
-            }
-
-            measurements.push_back(measurement);
-        }
-
-        return measurements;
-    }
-
-    int num_measurements;
-    std::vector<Vehicle> vehicles;
-};
 
 float pignistic_transformation(float free_mass, float occ_mass)
 {
@@ -375,7 +330,17 @@ int main(int argc, const char** argv)
 	laser_params.fov = 120.0f;
 	laser_params.max_range = 50.0f;
 
+	// Just to init cuda
+	cudaDeviceSynchronize();
+
+	auto begin = chrono::high_resolution_clock::now();
+
 	dogm::DOGM grid_map(params, laser_params);
+
+	auto end = chrono::high_resolution_clock::now();
+	auto dur = end - begin;
+	auto ms = std::chrono::duration_cast<std::chrono::milliseconds>(dur).count();
+	std::cout << "### DOGM initialization took: " << ms << " ms" << " ###" << std::endl << std::endl;
 
 	Simulator simulator(100);
 	simulator.addVehicle(Vehicle(6, glm::vec2(20, 10), glm::vec2(0, 0)));
@@ -392,14 +357,14 @@ int main(int argc, const char** argv)
 	{
 		grid_map.updateMeasurementGrid(sim_measurements[i].data(), sim_measurements[i].size());
 #endif
-		auto begin = chrono::high_resolution_clock::now();
+		begin = chrono::high_resolution_clock::now();
 
 		// Run Particle filter
 		grid_map.updateParticleFilter(delta_time);
 
-		auto end = chrono::high_resolution_clock::now();
-		auto dur = end - begin;
-		auto ms = std::chrono::duration_cast<std::chrono::milliseconds>(dur).count();
+		end = chrono::high_resolution_clock::now();
+		dur = end - begin;
+		ms = std::chrono::duration_cast<std::chrono::milliseconds>(dur).count();
 		std::cout << "### Iteration took: " << ms << " ms" << " ###" << std::endl;
 		std::cout << "######  Saving result  #######" << std::endl;
 		std::cout << "##############################" << std::endl;
