+++ dogm/src/dogm.cu
@@ -58,6 +58,16 @@ DOGM::DOGM(const GridParams& params, const LaserSensorParams& laser_params)
       particle_count(params.particle_count), grid_cell_count(grid_size * grid_size),
       new_born_particle_count(params.new_born_particle_count), block_dim(BLOCK_SIZE)
 {
+    int device;
+    CHECK_ERROR(cudaGetDevice(&device));
+
+    cudaDeviceProp device_prop;
+    CHECK_ERROR(cudaGetDeviceProperties(&device_prop, device));
+
+    int blocks_per_sm = device_prop.maxThreadsPerMultiProcessor / block_dim.x;
+    dim3 dim(device_prop.multiProcessorCount * blocks_per_sm);
+    particles_grid = birth_particles_grid = grid_map_grid = dim;
+
     CHECK_ERROR(cudaMallocManaged((void**)&grid_cell_array, grid_cell_count * sizeof(GridCell)));
     CHECK_ERROR(cudaMallocManaged((void**)&particle_array, particle_count * sizeof(Particle)));
     CHECK_ERROR(cudaMallocManaged((void**)&particle_array_next, particle_count * sizeof(Particle)));
@@ -71,19 +81,9 @@ DOGM::DOGM(const GridParams& params, const LaserSensorParams& laser_params)
     CHECK_ERROR(cudaMalloc(&birth_weight_array, new_born_particle_count * sizeof(float)));
     CHECK_ERROR(cudaMalloc(&born_masses_array, grid_cell_count * sizeof(float)));
 
-    CHECK_ERROR(cudaMalloc(&rng_states, particle_count * sizeof(curandState)));
+    CHECK_ERROR(cudaMalloc(&rng_states, particles_grid.x * block_dim.x * sizeof(curandState)));
 
     renderer = std::make_unique<Renderer>(grid_size, laser_params.fov, params.size, laser_params.max_range);
-
-    int device;
-    CHECK_ERROR(cudaGetDevice(&device));
-
-    cudaDeviceProp device_prop;
-    CHECK_ERROR(cudaGetDeviceProperties(&device_prop, device));
-
-    int blocks_per_sm = device_prop.maxThreadsPerMultiProcessor / block_dim.x;
-    dim3 dim(device_prop.multiProcessorCount * blocks_per_sm);
-    particles_grid = birth_particles_grid = grid_map_grid = dim;
 
     initialize();
 }
@@ -110,7 +110,7 @@ void DOGM::initialize()
     CHECK_ERROR(cudaStreamCreate(&particles_stream));
     CHECK_ERROR(cudaStreamCreate(&grid_stream));
 
-    setupRandomStatesKernel<<<particles_grid, block_dim>>>(rng_states, 123456, particle_count);
+    setupRandomStatesKernel<<<particles_grid, block_dim>>>(rng_states, 123456, particles_grid.x * block_dim.x);
 
     CHECK_ERROR(cudaGetLastError());
     CHECK_ERROR(cudaDeviceSynchronize());
